pfinet: Fix spurious EINTR errors from select
While sleeping in pthread_hurd_cond_timedwait_np, current->signal will be
overwritten by the management of other RPCs, so we have to give it some
value on wake up.

Also, if we previously got interrupted, we shouldn't ever try to wait again
until exiting from the RPC: an interrupt means we really want to try hard to
finish the RPC.

Thanks Zhaoming Luo <zhmingluo@163.com> for the deep investigation and
draft!