ext2fs: Always clean just-allocated inode
So that if we get a lazily-cleaned inode, we do not get disturbed by the
values, notably i_file_acl and i_dir_acl were not cleaned, and
i_block_group not checked.

b0ff48880bb40 ("Formerly ialloc.c.~6~") introduced only setting
dn_set_ctime when cleanup was missing, but we always want to set the
ctime for a new inode anyway.