nfs: Fix file creation for nfsv3
This adds new code to process the reply from a file creation operation that
applies equally to CREATE3, MKDIR3, SYMLINK3 and MKNOD3. RFC1813 permits
a reply without the created file's handle and it might be necessary to
perform a LOOKUP3 on the created file post-creation. Debian Linux NFSV3
server actually always returns the handle when compiling gnumach however
I have tested the case where it might not by adding code (now removed) to
pretend the handle wasn't present in the reply. I expect therefore that the
'skip_returned_stat' function that is added is rarely called.

Fixed netfs_attempt_mkdir() to process the NFSv3 reply properly.