nfs: Fix netfs_attempt_create_file for nfsv3
The v3 netfs_attempt_create_file() uses the 'exclusive' creation protocol
which cannot specify the file permission bits or uid/gid. These need to be
applied post-creation using SETATTR3. There was a missing call to do the chmod
part  which I've added. Previously the call to chown that was present did not
consider the return value. I have done so.