nfs: Implement statfs for V3.
There was little overlap with the v2 code so it got an entirely separate
implementation. There is the possibility of completing statfs.f_namelen
properly using FSINFO3 and PATHCONF3 rpc calls but I didn't know if
f_namelen should be the maximum length for a file name component or the
entire file name path.