nfs: generalise process_create_reply()
to allow it to be used for MKNOD and SYMLINK rpc calls. These file types
work differently to others in that a temporary file is created earlier and
its 'struct node' is migrated to the new file later.  A few adjustments to
the mutex locking to comply with the locking rules. A couple of mutex unlock
errors spotted: not being unlocked on malloc failure and an unnecessary
unlock in netfs_attempt_rename()