nfs: fix the handling of replies to MKNOD and SYMLINK
The Linux NFS server does not accept a setting of file size to 0 for
sockets and FIFOs so there is code to strip that for those cases. I also
spotted that the file modification times were not being applied to a
created file so there is now just one rpc to set all file attributes using
xdr_encode_create_state(). This sets the times to be server time which is
appropriate for a new file (even though it was created in the previous RPC
call). The previous code was supplying the 'gid' based on that returned by
the 'stat' but I don't think that can be relied on since the V3 protocol
states that CREATE3 using EXCLUSIVE cannot guarantee any file attributes
until after the SETATTR3.