lwip: allow setting an zero or broadcast address
After DHCPRELEASE, DHCP client sends a SIOCSIFADDR to set the
interface address to 0.0.0.0. This is not accepted by our IP
validator `ipv4config_is_valid()`. I think this use case is legit
so this commit makes the translator accept 0.0.0.0 as a valid address.
This assumes that a client trying to set IP to 0.0.0.0 is in fact trying
to remove the IP from the interface. Then it sets all fields to NONE, in
order to pass the validation.

* lwip/lwip-util.c: `configure_device`: Add the exception for 0.0.0.0
  and 255.255.255.255.
Message-ID: <20250729155244.7976-2-jlledom@mailfence.com>